requirements.txt
mkdocs-exclude==1.0.2
mkdocs-glightbox
mkdocs-llmstxt==0.3.1
mkdocs-macros-plugin==1.3.7
//copilot requirements//
chromadb
google-generativeai